Added some animations to try.
They look real bad, but the system works.
Need to come up with all poossible states for the thing.
What if you are standing up and loose a leg, you have to fall down from the point you are at... and so on.
There is no ragdoll or procedual animations to save you.