Fixed a bug
The animations still dont match. But there is crawling now.
How do we stand up if we have legs but no arms?
Also if you shoot a limb mid transition, loosing leg while standing up anim is running example, falling from standing looks a bit of...
Do we go ragdoll when we loose legs for a few sec... then turn to crawl if we have arms?