PlayerClimb Prototype
Very hastily thrown together code to try to evaluate if this should be a mechanic in the game. The player can transition between 2 different controllers, nonphys (normal) and climb. The player can transition by interacting with the vials in the scene. No combat in this test scene, it's simply to prove that this mechanic exists if we want to build on it

So yeah.. don't merge this to dev. It will probably break stuff!