Added EnemyDeathEvent and Listener
A GameObject with EnemyDeathEvent (needs Health-script)  will raise an event to the gameobject with EnemyDeathListener. In the event there is the position of where the enemy died.
The Listener is to be expanded upon as we need it to work with stuff...
Correct enemy prefabs needs the EnemyDeathEvent  attached. Lizards have it now