Occlusion on level2
added both manual and automatic occlusion on level two in order to increse performance. The namual does most of the lifting by accually disabling object so the automatic have fewer to sort and disable rendering.

Also body parts disabled as that gave 10+ frames increse. Too bad but they are using Skinned meshrenderer and was so spread out that it was very hard to get them into the manual occlusion system. Added 2 new prefabs for static bodyparts and scattered a few in the "dentist" area. yes room now have names like game callouts.

The spread of objects in the scene was moved and organized a bit.

New script to help with manuall occlusion was added and can be used similar to the existing trigger zone.

Level is playable, I tested it but not from main menu.

There is random elevator music when keypad is done also you can glitch out of the walls and do speedrun