Tried fixing light issues
VentShader is no longer used, new SpecialTriplanarShader is used instead as it is can be built. Used packed textures for metalness and smoothness, ie one texture with metal in read and smothness in green channels, blue and alpha is left blank. Also made a plugin for packing textures, bvery rudamentary version..

Lights are stupid and also dumb